+++ start/src/App.vue
@@ -1,3 +1,16 @@
 <script lang="ts" setup></script>
 
-<template></template>
+<template>
+  <f-space>
+    <f-button type="primary">主要按钮</f-button>
+    <f-button type="success">成功按钮</f-button>
+    <f-button type="warning">警告按钮</f-button>
+    <f-button type="warning">警告按钮</f-button>
+    <f-button type="warning">警告按钮</f-button>
+    <f-button type="warning">警告按钮</f-button>
+    <f-button type="warning">警告按钮</f-button>
+    <f-button type="warning">警告按钮</f-button>
+    <f-button type="warning">警告按钮</f-button>
+    <f-button type="warning">警告按钮</f-button>
+  </f-space>
+</template>