+++ packages/fighting-design/_hooks/use-form-check/index.ts
@@ -59,7 +59,6 @@ export const useFormCheck = (prop: FormProps): UseFormCheckReturn => {
     value: string,
     rules: FormItemRules | FormItemRulesItem
   ): string | boolean => {
-
     /**
      * 测试每一项规则
      *
@@ -105,7 +104,6 @@ export const useFormCheck = (prop: FormProps): UseFormCheckReturn => {
    */
   const validate = (): boolean => {
     getChildrenList.value.forEach((item: VNode): void => {
-
       /** 子组件规则 */
       const _rules: FormItemProps['rules'] = item.props && item.props.rules
       /** 子组件名字 */
@@ -113,18 +111,13 @@ export const useFormCheck = (prop: FormProps): UseFormCheckReturn => {
 
       /** 判断的每个自组件必须有 rules 和 name 参数 */
       if (item.props && _rules && _name && prop.model) {
-
         /** 检测父组件绑定的对象上是否存在子组件绑定的 name 属性 */
         if (_name in prop.model) {
-
           /** 获取需要检测的值 */
           const modelKeyVal: string = (prop.model as object)[_name as keyof object]
 
           /** 获取到规则校验的信息 */
-          const msg: string | boolean = checkRuleMassage(
-            modelKeyVal,
-            _rules
-          )
+          const msg: string | boolean = checkRuleMassage(modelKeyVal, _rules)
 
           childrenCheckResult[_name] = msg
         } else {