+++ packages/fighting-design/dialog/src/interface.ts
@@ -2,7 +2,7 @@ export type { DialogProps } from './props'
 
 /**
  * 回调方法
- * 
+ *
  * @param { Object } node 元素节点
  */
 export type DialogCallback = (node: Element) => void