+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `f-select` 组件 `width` 默认值改为 190
 - 新增 `f-trigger` 组件 `width` 配置项
 - 新增 `f-dropdown` 组件 `width` 配置项
+- 修复 `f-form-item` 组件 name 错误导致的报错
 
 ## 0.44.0 (2023-06-18)