+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - 新增 `f-trigger` 组件 `width` 配置项
 - 新增 `f-dropdown` 组件 `width` 配置项
 - 修复 `f-form-item` 组件 name 错误导致的报错
+- 优化 `f-space` 组件样式细节
 
 ## 0.44.0 (2023-06-18)